+++ packages/components-js/projects/jsdom-polyfill/src/normalizeCssNamespace.js
@@ -69,4 +69,3 @@ function normalizeCssNamespace(cssNamespace = globalThis.CSS) {
 }
 
 module.exports = { normalizeCssNamespace };
-