+++ packages/shared/src/testing/normalizeCssNamespace.ts
@@ -67,4 +67,3 @@ export const normalizeCssNamespace = (cssNamespace: object | undefined = globalT
     }
   }
 };
-